% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{add_footnote}
\alias{add_footnote}
\title{Add footnotes to Typst table}
\usage{
add_footnote(x, footnote)
}
\arguments{
\item{x}{A \code{typst_table} object.}

\item{footnote}{Character vector of footnotes to add to \code{x}}
}
\description{
This is equivalent to specifying the footnotes in the original call to
\code{ttab()}.
}
