% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{add_header}
\alias{add_header}
\title{Add header row above existing header}
\usage{
add_header(x, header)
}
\arguments{
\item{x}{A \code{typst_table} object.}

\item{header}{List or character vector of column headings to add. Must have one element per
column of x (after expansion of any multi-column cells). Must be a list (not a character
vector) if any elements contain custom formatting commands (e.g. \code{colspan()}).}
}
\description{
Similar to \code{kableExtra::add_header_above()}. Can be useful for grouped columns.
}
