% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{add_hline}
\alias{add_hline}
\title{Add horizontal lines in a table}
\usage{
add_hline(x, before)
}
\arguments{
\item{x}{A \code{typst_table} object.}

\item{before}{The table rows above which lines should be drawn.}
}
\description{
Lines span the full width of the table.
}
