% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{add_styling}
\alias{add_styling}
\title{Apply custom table styling}
\usage{
add_styling(x, style, replace = FALSE)
}
\arguments{
\item{x}{A \code{typst_table} or \code{typst_figure} object}

\item{style}{Character vector of Typst styling commands (or anything else)}

\item{replace}{Whether to replace any styles previously specified. Default is \code{FALSE} (new styles
are added to any previous specifications).}
}
\description{
Passed verbatim as Typst commands before the table itself
}
