% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{add_vline}
\alias{add_vline}
\title{Add vertical lines in a table}
\usage{
add_vline(x, before, start = 0, end = "none")
}
\arguments{
\item{x}{A \code{typst_table} object}

\item{before}{The table columns before which lines should be drawn}

\item{start, end}{The start and end rows of the line. Specified as in Typst's \verb{#table.vline()}
function.}
}
\description{
Add vertical lines in a table
}
