% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{cell_spec}
\alias{cell_spec}
\title{Apply formatting to table cells}
\usage{
cell_spec(x, row, col, italic = NULL, bold = NULL, align = NULL)
}
\arguments{
\item{x}{A \code{typst_table} object.}

\item{row, col}{Which cell to apply formatting to}

\item{italic, bold}{(optional) Logical scalars. Apply italic or bold
formatting.}

\item{align}{(optional) Align cell contents. Specified as in Typst's
\verb{#table.cell()}.}
}
\description{
More formatting options will be added in a later version. All formatting
arguments default to \code{NULL}, keeping the current cell formatting.
}
