% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMORprojects-package.R
\docType{package}
\name{CMORprojects-package}
\alias{CMORprojects}
\alias{CMORprojects-package}
\title{CMORprojects: Implement CMOR Data Analysis Project Workflow}
\description{
CMORprojects provides tools to optimise a research project data analysis workflow for research projects undertaken at CMOR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/uo-cmor/CMORprojects}
  \item Report bugs at \url{https://github.com/uo-cmor/CMORprojects/issues}
}

}
\author{
\strong{Maintainer}: Ross Wilson \email{ross.wilson@otago.ac.nz}

}
\keyword{internal}
