% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{col_spec}
\alias{col_spec}
\title{Apply formatting to table columns}
\usage{
col_spec(x, cols, italic = FALSE, bold = FALSE, header = TRUE)
}
\arguments{
\item{x}{A \code{typst_table} object.}

\item{cols}{Which columns to apply the formatting to}

\item{italic, bold}{(optional) Which formatting to apply. Defaults to none.}

\item{header}{Also apply formatting to header rows? Defaults to \code{TRUE}.}
}
\description{
More formatting options will be added in a later version.
}
