% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{colspan}
\alias{colspan}
\title{Define a multi-column cell}
\usage{
colspan(contents, n, align = "center", stroke = TRUE)
}
\arguments{
\item{contents}{Character string. The contents of the cell.}

\item{n}{The number of columns to span}

\item{align}{(optional) Alignment of the cell contents. Defaults to
\code{"center"}.}

\item{stroke}{(optional) Whether to draw a stroke under the cell. Defaults to
\code{TRUE}.}
}
\description{
Useful in the specification of grouped header rows. See \code{add_header()}.
}
