% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_json.R
\name{export_json}
\alias{export_json}
\title{Write outputs to JSON file for Typst import}
\usage{
export_json(
  r = list(),
  fig = list(),
  tbl = list(),
  path = "output/manuscript-inputs.json",
  auto_unbox = TRUE,
  ...
)
}
\arguments{
\item{r}{List of R objects}

\item{fig}{List of figure (SVG) files}

\item{tbl}{List of table (TYP) files}
}
\description{
Write outputs to JSON file for Typst import
}
