% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{landscape}
\alias{landscape}
\title{Output a table on its own landscape page}
\usage{
landscape(x)
}
\arguments{
\item{x}{A \code{typst_table} object.}
}
\description{
Useful for wide tables that don't fit in portrait.
}
