% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectsetup-package.R
\docType{package}
\name{projectsetup-package}
\alias{projectsetup}
\alias{projectsetup-package}
\title{projectsetup: Initialise CMOR Data Analysis Project Workflow}
\description{
projectsetup provides tools to initialise a research project data analysis workflow for research projects undertaken at CMOR.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rosswilson-nz/projectsetup}
  \item Report bugs at \url{https://github.com/rosswilson-nz/projectsetup/issues}
}

}
\author{
\strong{Maintainer}: Ross Wilson \email{ross.wilson@otago.ac.nz}

}
\keyword{internal}
