% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{row_spec}
\alias{row_spec}
\title{Apply formatting to table rows}
\usage{
row_spec(x, rows, italic = FALSE, bold = FALSE)
}
\arguments{
\item{x}{A \code{typst_table} object.}

\item{rows}{Which rows to apply the formatting to}

\item{italic, bold}{(optional) Which formatting to apply. Defaults to none.}
}
\description{
More formatting options will be added in a later version.
}
