% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{set_table_options}
\alias{set_table_options}
\title{Set Typst figure options}
\usage{
set_table_options(
  x,
  width,
  height,
  placement,
  caption,
  label,
  supplement,
  landscape
)
}
\arguments{
\item{x}{A Typst table (\code{CMORprojects_tbl})}

\item{placement}{Table placement. Either \code{"none"}, \code{"auto"}, \code{"top"},
\code{"bottom"}. The default is \code{"auto"}.}

\item{caption}{The table caption.}

\item{label}{Table label, used for cross-referencing}

\item{supplement}{Whether the table is to be placed in supplementary
material in the output. This only changes the Typst 'kind' parameter to
\code{"suppl-table"} instead of \code{"table"}. Typst templates may make use of
this to format the figure differently. Default is \code{FALSE.}}

\item{landscape}{Whether the table should be placed on its own landscape
page. Default is \code{FALSE}.}
}
\value{
A Typst table with the specified options set.
}
\description{
Set Typst figure options
}
