% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{span_cols}
\alias{span_cols}
\title{Make cells span multiple columns}
\usage{
span_cols(x, range, rows, align = "left", stroke = FALSE)
}
\arguments{
\item{x}{A \code{typst_table} object.}

\item{range}{The range (\verb{c(<min>, <max>)}) of columns to include in the span.}

\item{rows}{The rows in which to apply the span}

\item{align}{(optional) Alignment of the new multicolumn cell. Defaults to
\code{"left"}.}

\item{stroke}{(optional) Whether to draw a stroke under the merged cells.
Defaults to \code{FALSE}.}
}
\description{
If any cells to the right of the first one are non-empty, they will be
overwritten with a warning.
}
