% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{span_rows}
\alias{span_rows}
\title{Make cells span multiple rows}
\usage{
span_rows(x, range, cols, align = "top")
}
\arguments{
\item{x}{A \code{typst_table} object.}

\item{range}{The range (\verb{c(<min>, <max>)}) of rows to include in the span.}

\item{cols}{The columns in which to apply the span}

\item{align}{(optional) Alignment of the new multirow cell. Defaults to
\code{"top"}.}
}
\description{
If any cells below the top one are non-empty, they will be overwritten with a
warning.
}
