% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{supplement}
\alias{supplement}
\title{Mark a table as supplementary material}
\usage{
supplement(x)
}
\arguments{
\item{x}{A \code{typst_table} object.}
}
\description{
Just specifies the Typst \code{kind} as \code{"suppl-table"} instead of \code{"table"}. The Typst template use
needs to do something with this to have any effect.
}
