% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-figures.R
\name{tfig}
\alias{tfig}
\title{Create figures in Typst format}
\usage{
tfig(
  x,
  caption = NULL,
  label = NULL,
  placement = NULL,
  width = NULL,
  height = NULL,
  footnotes = NULL
)
}
\arguments{
\item{x}{Path to a saved image in SVG, PNG, JPEG, or GIF format.}

\item{caption}{The figure caption.}

\item{label}{Figure label, used for cross-referencing}

\item{placement}{(optional) Figure placement. As in Typst's #figure() function.}

\item{width, height}{(optional) Image width and height. As in Typst's #image() function.}

\item{footnotes}{Footnotes to add below the table. Pass a vector for multiple
footnotes. At this stage, footnote numbering needs to be added manually
(as does the corresponding numbering in table cells).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated in favour of the new versions provided by the
\code{ttables} package.

This wraps the path to a saved image in a Typst figure environment with caption and label.
}
