% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ttab}
\alias{ttab}
\title{Create tables in Typst format}
\usage{
ttab(
  x,
  caption = NULL,
  label = NULL,
  align = "left",
  widths = "auto",
  placement = NULL,
  footnotes = NULL
)
}
\arguments{
\item{x}{An R data frame.}

\item{caption}{The table caption.}

\item{label}{Table label, used for cross-referencing}

\item{align}{Column alignment. Either a single value or a vector with the
same length as the number of columns in \code{x}. Possible values are
\code{"left"},  \code{"center"}, \code{"right"}.}

\item{widths}{Columns widths. Either 'auto' for automatically determined
column widths, a vector of numeric values (for relative widths), or a
vector of Typst track sizes ('auto' or Typst fixed, relative, or
fractional lengths).}

\item{placement}{(optional) Table placement. As in Typst's #figure()
function.}

\item{footnotes}{Footnotes to add below the table. Pass a vector for multiple
footnotes. At this stage, footnote numbering needs to be added manually
(as does the corresponding numbering in table cells).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated in favour of the new versions provided by the
\code{ttables} package.

This is a table generator inspired by the \emph{kableExtra} package. This base
function takes an R data frame and sets up a basic Typst table. Auxiliary
functions allow the adding of additional formatting features in a manner
similar to that of the \emph{kableExtra} formatting functions.
}
\keyword{internal}
