% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-figures.R
\name{typst_figure}
\alias{typst_figure}
\title{Create figures in Typst format}
\usage{
typst_figure(
  x,
  caption = NULL,
  label = NULL,
  placement = "auto",
  width = "auto",
  height = "auto",
  footnotes = NULL,
  supplement = FALSE
)
}
\arguments{
\item{x}{Path to a saved image in SVG, PNG, JPEG, or GIF format.}

\item{caption}{The figure caption.}

\item{label}{Figure label, used for cross-referencing}

\item{placement}{(optional) Figure placement. As in Typst's #figure() function.}

\item{width, height}{(optional) Image width and height. As in Typst's #image() function.}

\item{footnotes}{Footnotes to add below the table.}
}
\description{
This wraps the path to a saved image in a Typst figure environment with caption and label.
}
