% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{typst_table}
\alias{typst_table}
\title{Create tables in Typst format}
\usage{
typst_table(
  x,
  caption = NULL,
  label = NULL,
  placement = "auto",
  supplement = FALSE
)
}
\arguments{
\item{x}{Path to a saved table in TYP format.}

\item{caption}{The table caption.}

\item{label}{Table label, used for cross-referencing}

\item{placement}{(optional) Table placement. As in Typst's #figure()
function.}

\item{supplement}{Whether the table should be marked as supplementary
material.}
}
\description{
This wraps the path to a saved Typst table with
caption and label.
}
