% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-projects_author.R
\name{vec_ptype_abbr.projects_author}
\alias{vec_ptype_abbr.projects_author}
\title{abreviation for  project author}
\usage{
\method{vec_ptype_abbr}{projects_author}(x, ...)
}
\arguments{
\item{x}{something i guess}

\item{...}{else}
}
\description{
abreviation for  project author
}
