% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-projects_stage.R
\name{vec_ptype_abbr.projects_stage}
\alias{vec_ptype_abbr.projects_stage}
\title{abreviation for  project stage}
\usage{
\method{vec_ptype_abbr}{projects_stage}(x, ...)
}
\arguments{
\item{x}{something i guess}

\item{...}{else}
}
\description{
abreviation for  project stage
}
