% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst-tables.R
\name{col_spec}
\alias{col_spec}
\title{Apply formatting to table columns}
\usage{
col_spec(x, cols, italic = NULL, bold = NULL, align = NULL, header = TRUE)
}
\arguments{
\item{x}{A \code{typst_table} object.}

\item{cols}{Which columns to apply the formatting to}

\item{italic, bold}{(optional) Logical scalars. Apply italic or bold
formatting.}

\item{align}{(optional) Align cell contents. Specified as in Typst's
\verb{#table.cell()}.}

\item{header}{Also apply formatting to header rows? Defaults to \code{TRUE}.}
}
\description{
More formatting options will be added in a later version. All formatting
arguments default to \code{NULL}, keeping the current cell formatting.
}
