% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_json.R
\name{export_json}
\alias{export_json}
\title{Write outputs to JSON file for Typst import}
\usage{
export_json(x, path = "output/values.json", auto_unbox = TRUE, ...)
}
\arguments{
\item{x}{List of R objects}

\item{path}{JSON file to write \code{x} to}

\item{auto_unbox, ...}{Passed to \code{jsonlite::write_json()}}
}
\description{
Write outputs to JSON file for Typst import
}
